pyproject.toml
[tool.poetry]
name = "auto_journalist"
version = "0.1.0"
description = "Auto-Journalist: A multi-agent daily news briefing generator"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
openai = "^0.27.0"
sqlalchemy = "^2.0"
alembic = "^1.10.0"
psycopg2-binary = "^2.9"
requests = "^2.28"
jinja2 = "^3.1"
markdown-it-py = "^2.2"
python-dotenv = "^1.0"
newsapi-python = "^0.2"
guardian = "^0.2.3"
sendgrid = "^6.9"
apscheduler = "^3.10"
pytest = {extras = ["cov"], version = "^7.0"}
vcrpy = "^4.0"

[tool.poetry.dev-dependencies]
black = "^23.3.0"
ruff = "^0.0"
mypy = "^1.4"
pytest = "^7.0"
pytest-asyncio = "^0.21"
pytest-approved = "^1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

